Refactor: Service method 잘못된 위치 변경
