+++ backend/src/routers/v1/posts.ts
@@ -1,19 +1,19 @@
 import { Request, Response } from 'express';
-import { JsonController, Req, Res, Post, Delete } from 'routing-controllers';
+import { JsonController, Req, Post, Delete, Get } from 'routing-controllers';
 
 import { PostService, CommentService } from 'src/services';
 
 @JsonController('/posts')
 export default class PostsRouter {
   @Post('/')
-  async postPost(@Req() request: Request, @Res() response: Response) {
+  async postPost(@Req() request: Request) {
     const data = request.body;
     const result = await PostService.createPost(data);
     return result;
   }
 
   @Post('/:postId/comments')
-  async postComment(@Req() request: Request, @Res() response: Response) {
+  async postComment(@Req() request: Request) {
     const { postId } = request.params;
     const data = request.body;
     const result = await CommentService.createComment(data, postId);
@@ -21,7 +21,7 @@ export default class PostsRouter {
   }
 
   @Post('/:postId/comments/:commentId/likes')
-  async postCommentLike(@Req() request: Request, @Res() response: Response) {
+  async postCommentLike(@Req() request: Request) {
     const { postId, commentId } = request.params;
     const data = request.body;
     const result = await CommentService.createCommentLike(data.userID, postId, commentId);
@@ -29,17 +29,25 @@ export default class PostsRouter {
   }
 
   @Post('/:postId/likes')
-  async postPostLike(@Req() request: Request, @Res() response: Response) {
+  async postPostLike(@Req() request: Request) {
     const { postId } = request.params;
     const data = request.body;
-    const result = await CommentService.createPostLike(data.userID, postId);
+    const result = await PostService.createPostLike(data.userID, postId);
     return result;
   }
 
   @Delete('/:postId/likes/:likeId')
-  async deletePostLike(@Req() request: Request, @Res() response: Response) {
+  async deletePostLike(@Req() request: Request) {
     const { postId, likeId } = request.params;
-    const result = await CommentService.removePostLike(postId, likeId);
+    const result = await PostService.removePostLike(postId, likeId);
+    return result;
+  }
+
+  @Get('posts')
+  async getRandomPost(@Req() request: Request) {
+    const { type } = request.query;
+    let result;
+    if (type === 'random') result = PostService;
     return result;
   }
 }
